Greatly improved how analysis is done.

+++ Source/BalanceSpecsGUI/DataContexts/SingleRunResultsContext.cs
@@ -91,7 +91,7 @@ namespace BalanceSpecsGUI.DataContexts
         private void ThreadCode()
         {
             Running = true;
-            Results = GeneticAlgorithm.GeneticAlgorithm.BalanceGA.RunGamesRemote(JsonParams, Vector, RandSeed);
+            Results = GeneticAlgorithm.GeneticAlgorithm.BalanceGA.RunGamesRemote(JsonParams, Vector, (int)RandSeed);
             Running = false;
         }
